requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attr==0.3.1
azure-core==1.29.5
azure-storage-blob==12.18.3
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1695989934239/work
certifi==2023.7.22
cffi==1.16.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1698833585322/work
click==8.1.7
colorama==0.4.6
contourpy @ file:///Users/runner/miniforge3/conda-bld/contourpy_1699041448398/work
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1696677705766/work
defusedxml==0.7.1
Django==3.2.20
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
expiringdict==1.2.2
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1698714947081/work
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1699023568720/work
fsspec==2023.10.0
gmpy2 @ file:///Users/runner/miniforge3/conda-bld/gmpy2_1666808749046/work
google-api-core==2.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio-status==1.59.2
htmlmin==0.1.12
huggingface-hub==0.18.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
ijson==3.2.3
inflection==0.5.1
isodate==0.6.1
itypes==1.2.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
jmespath==0.10.0
joblib==1.3.2
jsonschema==3.2.0
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/kiwisolver_1695380058985/work
label-studio==1.8.2.post1
label-studio-converter==0.0.54rc0
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
lockfile==0.12.2
lxml==4.9.3
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1695367660391/work
matplotlib @ file:///Users/runner/miniforge3/conda-bld/matplotlib-suite_1698868590489/work
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1678228039184/work
munkres==1.1.4
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1698504735452/work
nltk==3.6.7
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1694920094885/work/dist/numpy-1.26.0-cp311-cp311-macosx_11_0_arm64.whl#sha256=6909902123b8421906e90ad77fb0041d9eb2d95bbdc29f3d09c7d244b0e0e5a5
openapi-codec==1.3.2
ordered-set==4.0.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.2
Pillow @ file:///Users/runner/miniforge3/conda-bld/pillow_1697423665652/work
proto-plus==1.22.3
psycopg2-binary==2.9.6
pycparser==2.21
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1690737849915/work
pyRFC3339==1.1
pyrsistent==0.20.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-json-logger==2.0.4
pytz==2023.3.post1
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1695373486380/work
redis==3.5.3
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
rq==1.10.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
rules==2.2
s3transfer==0.3.7
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
semver==2.13.0
sentry-sdk==1.34.0
simplejson==3.19.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sqlparse==0.4.4
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1684180540116/work
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
timm==0.9.10
torch==2.1.0
torchaudio==2.1.0
torchsampler==0.1.2
torchvision==0.16.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1695373481350/work
tqdm==4.66.1
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
ua-parser==0.18.0
ujson==5.8.0
uritemplate==4.1.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1697720414277/work
user-agents==2.2.0
webencodings==0.5.1
xmljson==0.2.0
